Enum tiny_frame::Checksum

pub enum Checksum {
    None,
    Xor,
    Crc16,
    Crc32,
}

Checksum types.

Variants
None

The no-op checksum type.

Xor

This checksum type will return the bit-flipped xor of all the bytes in the buffer.

Crc16

This checksum type will return the CRC16 sum with the polynomial 0x8005 (x^16 + x^15 + x^2 + 1).

Crc32

This checksum type will return the CRC32 sum with the polynomial 0xedb88320.

Methods

impl Checksum

pub fn sum(&self, buffer: &[u8]) -> u32

Calculates the checksum of the given buffer.

Examples

let buffer: Vec<u8> = vec![1, 2, 3, 4, 5];
assert_eq!(Checksum::Xor.sum(&buffer) as u8, !(1 ^ 2 ^ 3 ^ 4 ^ 5));

pub fn append_sum(&self, buffer: &mut Vec<u8>)

Appends the checksum of the given buffer to itself.

Examples

let mut buffer: Vec<u8> = vec![1, 2, 3, 4, 5];
Checksum::Xor.append_sum(&mut buffer);
assert_eq!(buffer, vec![1, 2, 3, 4, 5, !(1 ^ 2 ^ 3 ^ 4 ^ 5)]);