Trait tiny_frame::number::BufferWritable

pub trait BufferWritable {
    fn write_to_buf<W: Write>(&self, buf: &mut W) -> Result<()>;
}

A number type that can be written to a buffer using big endian encoding.

Required Methods

fn write_to_buf<W: Write>(&self, buf: &mut W) -> Result<()>

Appends the big endian byte value to the buffer.

Examples

let mut buffer: Vec<u8> = Vec::new();
12u8.write_to_buf(&mut buffer); // 12
280u16.write_to_buf(&mut buffer); // 1 * 256 + 24
assert_eq!(buffer, vec![12, 1, 24]);