Trait tiny_frame::BufferWritable

pub trait BufferWritable {
    fn write_to_buf(&self, buf: &mut Vec<u8>);
}

A number type that can be written to a buffer using big endian encoding.

Required Methods

fn write_to_buf(&self, buf: &mut Vec<u8>)

Appends the big endian byte value to the buffer.

Examples

let mut buffer: Vec<u8> = Vec::new();
12u8.write_to_buf(&mut buffer); // 12
280u16.write_to_buf(&mut buffer); // 1 * 256 + 24
assert_eq!(buffer, vec![12, 1, 24]);