Struct tiny_frame::TinyFrame

pub struct TinyFrame<Len, ID, Type> {
    pub peer_bit: Peer,
    pub parser_timeout: Option<Ticks>,
    pub sof_byte: Option<u8>,
    pub chunk_size: usize,
    pub cksum: Checksum,
    pub write: Option<Box<Fn(&mut TinyFrame<Len, ID, Type>, &[u8])>>,
    pub claim_tx: Option<Box<Fn(&TinyFrame<Len, ID, Type>)>>,
    pub release_tx: Option<Box<Fn(&TinyFrame<Len, ID, Type>)>>,
    // some fields omitted
}

A TinyFrame instance.

Len is the length field type, ID is the ID field type, and Type is the type field type. You may use custom types, but in general you should pick from u8, u16, u32, and u64.

Fields
peer_bit: Peer

The peer bit.

parser_timeout: Option<Ticks>

The parser timeout after which the parser will be reset.

sof_byte: Option<u8>

The optional start-of-header byte.

chunk_size: usize

The chunk size. 1024 by default.

cksum: Checksum

The checksum type. Xor by default.

write: Option<Box<Fn(&mut TinyFrame<Len, ID, Type>, &[u8])>>

A function called every time something is written. This must be implemented.

claim_tx: Option<Box<Fn(&TinyFrame<Len, ID, Type>)>>

A function called before writing, for claiming the TX interface.

release_tx: Option<Box<Fn(&TinyFrame<Len, ID, Type>)>>

A function called after writing, for releasing the TX interface.

Methods

impl<Len, ID, Type> TinyFrame<Len, ID, Type> where
Len: GenericNumber,
ID: GenericNumber,
Type: GenericNumber,

pub fn new(peer_bit: Peer) -> TinyFrame<Len, ID, Type>

Creates a new TinyFrame with the specified peer bit.

pub fn reset_parser(&mut self)

Resets the parser.

pub fn add_id_listener(
&mut self,
id: ID,
cb: Box<Listener<Len, ID, Type>>,
timeout: Option<Ticks>
) -> Rc<IDListener<Len, ID, Type>>

Adds an ID listener.

The listener will be called if a message with the specified ID is received. If timeout is not None, the listener will expire after the specified number of ticks.

Note that if the returned IDListener is dropped, the listener is too.

pub fn add_type_listener(
&mut self,
msg_type: Type,
cb: Box<Listener<Len, ID, Type>>
) -> Rc<TypeListener<Len, ID, Type>>

Adds a type listener.

The listener will be called if a message with the specified type is received.

Note that if the returned TypeListener is dropped, the listener is too.

pub fn add_generic_listener(
&mut self,
cb: Box<Listener<Len, ID, Type>>
) -> Rc<GenericListener<Len, ID, Type>>

Adds a generic listener.

The listener will be called every time a message is received.

Note that if the returned GenericListener is dropped, the listener is too.

pub fn send(&mut self, msg: Msg<ID, Type>)

Sends a message.

pub fn query(
&mut self,
msg: Msg<ID, Type>,
listener: Box<Listener<Len, ID, Type>>,
timeout: Option<Ticks>
) -> Rc<IDListener<Len, ID, Type>>

Sends a message and binds an ID listener to listen for the response.

Note that if the returned IDListener is dropped, the listener is too.

pub fn respond(&mut self, msg: Msg<ID, Type>)

Sends a response.

pub fn accept(&mut self, buffer: &Vec<u8>)

Reads a buffer. This is syntax sugar for accept_byte.

pub fn accept_byte(&mut self, byte: u8)

Reads one byte.

pub fn tick(&mut self)

This function should be called periodically to time-out partial frames and ID listeners.